ti-sci-fw: backport recipe from meta-ti dunfell

+++ recipes-bsp/ti-sci-fw/ti-sci-fw_2020.02.bb
@@ -8,30 +8,41 @@ DEPENDS = "openssl-native u-boot-mkimage-native dtc-native"
 CLEANBROKEN = "1"
 PR = "r0"
 
-COMPATIBLE_MACHINE = "tqma65xx"
+# Loaded by R5F core
+COMPATIBLE_MACHINE = "k3r5"
+COMPATIBLE_MACHINE_aarch64 = "null"
 
 PACKAGE_ARCH = "${MACHINE_ARCH}"
 
 TI_SECURE_DEV_PKG ?= ""
 export TI_SECURE_DEV_PKG
 
+SRCREV = "c8decf64be551dfd1244cd1d231a97eb2255fb80"
+BRANCH ?= "ti-linux-firmware"
 SRCREV_imggen = "c0d6e6ebc85d9e7e4e02a7e6364cadf31c6fad0d"
 SRCREV_FORMAT = "imggen"
 
 SRC_URI = " \
-        git://git.ti.com/k3-image-gen/k3-image-gen.git;protocol=git;branch=master;destsuffix=imggen;name=imggen \
+	git://git.ti.com/processor-firmware/ti-linux-firmware.git;protocol=git;branch=${BRANCH} \
+	git://git.ti.com/k3-image-gen/k3-image-gen.git;protocol=git;branch=master;destsuffix=imggen;name=imggen \
 "
 
 S = "${WORKDIR}/git"
 
-SYSFW_SOC = "am65x_sr2"
-SYSFW_CONFIG = "evm"
+SYSFW_SOC ?= "unknown"
+SYSFW_CONFIG ?= "unknown"
 
 SYSFW_PREFIX = "ti-sci-firmware"
+SYSFW_SUFFIX ?= "unknown"
+
+SYSFW_BASE = "${SYSFW_PREFIX}-${SYSFW_SOC}-${SYSFW_SUFFIX}"
+
+SYSFW_TISCI = "${S}/ti-sysfw/${SYSFW_BASE}*.bin"
 
 SYSFW_BINARY = "sysfw-${SYSFW_SOC}-${SYSFW_CONFIG}.itb"
-SYSFW_IMAGE = "sysfw-${PV}.itb"
-SYSFW_SYMLINK = "sysfw.itb"
+SYSFW_VBINARY = "sysfw-${PV}-${SYSFW_SOC}-${SYSFW_CONFIG}.itb"
+SYSFW_IMAGE = "sysfw-${SYSFW_SOC}-${SYSFW_CONFIG}.itb"
+SYSFW_SYMLINK ?= "sysfw.itb"
 
 CFLAGS[unexport] = "1"
 LDFLAGS[unexport] = "1"
@@ -39,21 +50,27 @@ AS[unexport] = "1"
 LD[unexport] = "1"
 
 do_configure[noexec] = "1"
-do_populate_lic[noexec] = "1"
 
 EXTRA_OEMAKE = "\
-    CROSS_COMPILE="${TARGET_PREFIX}" SOC="${SYSFW_SOC}"\
- \
+    CROSS_COMPILE=${TARGET_PREFIX} SYSFW_DL_URL='' SYSFW_PATH="${SYSFW_TISCI}" SOC=${SYSFW_SOC} CONFIG=${SYSFW_CONFIG} \
 "
+EXTRA_OEMAKE_HS = " \
+    HS=1 SYSFW_HS_PATH="${S}/ti-sysfw/${SYSFW_BASE}-enc.bin" SYSFW_HS_INNER_CERT_PATH="${S}/ti-sysfw/${SYSFW_BASE}-cert.bin" \
+"
+EXTRA_OEMAKE_append = "${@['',' ${EXTRA_OEMAKE_HS}']['${SYSFW_SUFFIX}' == 'hs']}"
+
 do_compile() {
-        cd ${WORKDIR}/imggen/
-        oe_runmake
+	cd ${WORKDIR}/imggen/
+	oe_runmake
 }
 
 do_install() {
-        install -d ${D}/boot
-        install -m 644 ${WORKDIR}/imggen/${SYSFW_BINARY} ${D}/boot/${SYSFW_IMAGE}
-        ln -sf ${SYSFW_IMAGE} ${D}/boot/${SYSFW_SYMLINK}
+	install -d ${D}/boot
+	install -m 644 ${WORKDIR}/imggen/${SYSFW_BINARY} ${D}/boot/${SYSFW_VBINARY}
+	ln -sf ${SYSFW_VBINARY} ${D}/boot/${SYSFW_IMAGE}
+	if [ ! -z "${SYSFW_SYMLINK}" ]; then
+		ln -sf ${SYSFW_VBINARY} ${D}/boot/${SYSFW_SYMLINK}
+	fi
 }
 
 FILES_${PN} = "/boot"
@@ -61,10 +78,16 @@ FILES_${PN} = "/boot"
 inherit deploy
 
 do_deploy () {
-        install -d ${DEPLOYDIR}
-        install -m 644 ${WORKDIR}/imggen/${SYSFW_BINARY} ${DEPLOYDIR}/${SYSFW_IMAGE}
-        rm -f ${DEPLOYDIR}/${SYSFW_SYMLINK}
-        ln -sf ${SYSFW_IMAGE} ${DEPLOYDIR}/${SYSFW_SYMLINK}
+	install -d ${DEPLOYDIR}
+	install -m 644 ${WORKDIR}/imggen/${SYSFW_BINARY} ${DEPLOYDIR}/${SYSFW_VBINARY}
+	rm -f ${DEPLOYDIR}/${SYSFW_IMAGE}
+	ln -sf ${SYSFW_VBINARY} ${DEPLOYDIR}/${SYSFW_IMAGE}
+	if [ ! -z "${SYSFW_SYMLINK}" ]; then
+		rm -f ${DEPLOYDIR}/${SYSFW_SYMLINK}
+		ln -sf ${SYSFW_VBINARY} ${DEPLOYDIR}/${SYSFW_SYMLINK}
+	fi
+
+	install -m 644 ${SYSFW_TISCI} ${DEPLOYDIR}/
 }
 
 addtask deploy before do_build after do_compile
